dev: Port Maven build config to SBT
It doesn't do Javascript minification and token replacement yet, but
looks very usable for development.

For example, I can do `sbt console` and it will drop me into the scala
interactive shell with all the project dependencies on the classpath.

Or, `~;container:start; container:reload /` will start the app in an
embedded jetty and incrementally compile and restart every time I make
a change in a source file.